requirements.txt
cffi==1.17.1
cryptography==43.0.3
mysql-connector-python==9.2.0
pycparser==2.22
PyMySQL==1.1.1
